+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Changelog
 - **Fix**: Fix support for Anvil's `exclude` argument in dependency graph annotations when Anvil interop is enabled.
 - **Fix**: Fix colliding reports destinations in KMP projects by disambiguating with the target path as another subdir indirection.
 - **Fix**: Fix qualifiers not being propagated to generated `MembersInjector` classes for multibindings.
+- **Fix**: Remove misleading "doesn't appear to be visible to this compilation" hint, as this hint was misreporting.
 - [gradle] Add `generateMetroGraphMetadata` task that writes a merged JSON dump of all binding graphs in the project. This can be chained from the `GenerateGraphMetadataTask` for further processing.
 - [gradle] Add `MetroArtifacts` API for accessing Metro reports and graph metadata directories.
 - [gradle] Mark `metro.reportsDestination` as delicate/opt-in.